cscvs to tla changeset 1481
Author: franl
Date: 2005/01/20 02:07:44
(erc-button-add-face): The face added by this function is more important than
the existing text's face, so we now prepend erc-button-face to the list of
existing faces when adding a button.  To instead append erc-button-face to
existing faces, set variable `erc-button-face-has-priority' to nil.
(erc-button-press-button): Silenced a byte-compiler warning about a too few
arguments in a call to `error'.
git-archimport-id: mwolson@gnu.org--2005/erc--cvs-head--0--patch-6