Get rid of most CVS-generated version strings.
All erc-*-version strings are defined to be the same as
`erc-version-string'.  I think it's unnecessary to have a version for
each module.  Moreover, it's a royal pain in the neck to keep them
up-to-date with the CVS version, so to heck with it.  I won't (yet) apply
this to CVS head or CVS 5.0 release branch, but the Arch-managed versions
should definitely have this done.
git-archimport-id: mwolson@gnu.org--2005/erc--cvs-head--0--patch-52